backend/requirements/base.txt
Django>=5.0
djangorestframework>=3.15
bcrypt==4.1.3
django-cors-headers==4.4.0

pymongo>=4.7            # Mongo driver
python-dotenv>=1.0      # load .env
PyJWT>=2.9              # JWT auth
passlib[bcrypt]>=1.7    # password hashing

rapidfuzz>=3.9          # fuzzy matching for queries
# optional (add later for smarter search)
sentence-transformers>=3.0
numpy>=1.26
